requirements.txt
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.0
PyJWT==2.3.0
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3

